+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/user_services_functions/start_user_services.go
@@ -86,20 +86,20 @@ func StartUserServices(
 		return nil, nil, stacktrace.Propagate(err, "An error occurred registering services with IDs '%v'", serviceIDsToRegister)
 	}
 	// Defer an undo to all the successful registrations in case an error occurs in future phases
-	serviceIDsToRemove := map[service.ServiceID]bool{}
-	for serviceID, _ := range successfulRegistrations {
-		serviceIDsToRemove[serviceID] = true
+	serviceGUIDsToRemove := map[service.ServiceGUID]bool{}
+	for _, registration := range successfulRegistrations {
+		serviceGUIDsToRemove[registration.GetGUID()] = true
 	}
 	defer func() {
-		if len(serviceIDsToRemove) == 0 {
+		if len(serviceGUIDsToRemove) == 0 {
 			return
 		}
 		userServiceFilters := &service.ServiceFilters{
-			IDs: serviceIDsToRemove,
+			GUIDs: serviceGUIDsToRemove,
 		}
 		_, failedToDestroyGUIDs, err := DestroyUserServices(ctx, enclaveID, userServiceFilters, cliModeArgs, apiContainerModeArgs, engineServerModeArgs, kubernetesManager)
 		if err != nil {
-			logrus.Errorf("Attempted to destroy all services with IDs '%v' together but had no success. You must manually destroy the services! The following error had occurred:\n'%v'", serviceIDsToRemove, err)
+			logrus.Errorf("Attempted to destroy all services with GUIDs '%v' together but had no success. You must manually destroy the services! The following error had occurred:\n'%v'", serviceGUIDsToRemove, err)
 			return
 		}
 		if len(failedToDestroyGUIDs) == 0 {
@@ -143,14 +143,16 @@ func StartUserServices(
 		}
 	}
 
-	// Get existing objects
-	serviceIDsToFilter := map[service.ServiceID]bool{}
-	for serviceID := range serviceConfigsToStart {
-		serviceIDsToFilter[serviceID] = true
+	// Get existing objects by GUID. We use GUIDs and not IDs as IDs can match services being deleted or services in other Enclaves.
+	serviceGUIDsToFilter := map[service.ServiceGUID]bool{}
+	for _, registration := range successfulRegistrations {
+		guid := registration.GetGUID()
+		serviceGUIDsToFilter[guid] = true
 	}
 	existingObjectsAndResourcesFilters := &service.ServiceFilters{
-		IDs: serviceIDsToFilter,
+		GUIDs: serviceGUIDsToFilter,
 	}
+	// This is safe, as there's an N -> 1 mapping between GUID and ID and the GUIDs that we filter on don't have any matching IDs
 	existingObjectsAndResources, err := shared_helpers.GetMatchingUserServiceObjectsAndKubernetesResourcesByServiceID(
 		ctx,
 		enclaveID,
@@ -161,15 +163,15 @@ func StartUserServices(
 		kubernetesManager,
 	)
 	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred getting user service objects and Kubernetes resources matching service IDs '%v'", serviceIDsToFilter)
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting user service objects and Kubernetes resources matching service GUIDs '%v'", serviceGUIDsToFilter)
 	}
-	for serviceID, _ := range serviceIDsToFilter {
+	for serviceID, _ := range successfulRegistrations {
 		if _, found := existingObjectsAndResources[serviceID]; !found {
 			failedServicesPool[serviceID] = stacktrace.NewError("Couldn't find any service registrations for service ID '%v'. This is a bug in Kurtosis.", serviceID)
 			delete(serviceConfigsToStart, serviceID)
 		}
 	}
-	if len(existingObjectsAndResources) > len(serviceIDsToFilter) {
+	if len(existingObjectsAndResources) > len(serviceGUIDsToFilter) {
 		// Should never happen because service GUIDs should be unique
 		return nil, nil, stacktrace.NewError("Found more than one service registration matching service GUIDs; this is a bug in Kurtosis")
 	}
@@ -194,8 +196,9 @@ func StartUserServices(
 	}
 
 	// Do not remove services that were started successfully
-	for serviceID, _ := range successfulServicesPool {
-		delete(serviceIDsToRemove, serviceID)
+	for _, service := range successfulServicesPool {
+		guid := service.GetRegistration().GetGUID()
+		delete(serviceGUIDsToRemove, guid)
 	}
 	logrus.Debugf("Started services '%v' succesfully while '%v' failed", successfulServicesPool, failedServicesPool)
 	return successfulServicesPool, failedServicesPool, nil