+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+### Fixes
+* Fixed a bug where services wouldn't start in `KubernetesBackend.StartUserServices` as we'd get older GUIDs or GUIDs from other enclaves that had been associated with the ServiceID earlier.
+* Fixed a bug where we are destroying services why `ID` which is dangerous (as we may select across enclaves), now destroying by `GUID` instead.
+
 # 0.38.0
 ### Features
 * Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute